+++ Chapter.Net.WPF.Converters/EqualsToBooleanConverter/EqualsToBooleanConverter.cs
@@ -5,6 +5,7 @@
 // -----------------------------------------------------------------------------------------------------------------
 
 using System;
+using System.ComponentModel;
 using System.Globalization;
 using System.Windows.Data;
 
@@ -13,28 +14,43 @@ using System.Windows.Data;
 namespace Chapter.Net.WPF.Converters;
 
 /// <summary>
-///     Equals two values and returns its result.
+///     Compares a single value to a variable or a list of values to each other and returns its boolean representation.
 /// </summary>
-[ValueConversion(typeof(object), typeof(bool), ParameterType = typeof(object))]
+[ValueConversion(typeof(object), typeof(bool))]
 public class EqualsToBooleanConverter : SingleAndMultiValueConverter
 {
     /// <summary>
-    ///     The value to return if equals is true; otherwise the opposite.
+    ///     The return if equals is true.
     /// </summary>
     /// <value>Default: true.</value>
-    public bool IsEqual { get; set; } = true;
+    [DefaultValue(true)]
+    public bool? IsEqual { get; set; } = true;
 
     /// <summary>
-    ///     Equals two values and returns its result.
+    ///     The return if equals is false.
     /// </summary>
-    /// <param name="value">The first value.</param>
+    /// <value>Default: false.</value>
+    [DefaultValue(false)]
+    public bool? IsNotEqual { get; set; } = false;
+
+    /// <summary>
+    ///     The value to compare with.
+    /// </summary>
+    /// <value>Default: null.</value>
+    [DefaultValue(null)]
+    public object CompareWith { get; set; } = null;
+
+    /// <summary>
+    ///     Compares a single value to a variable and returns its boolean representation.
+    /// </summary>
+    /// <param name="value">The value to convert.</param>
     /// <param name="targetType">Unused.</param>
-    /// <param name="parameter">The second value.</param>
+    /// <param name="parameter">Unused.</param>
     /// <param name="culture">Unused.</param>
-    /// <returns>The boolean representation of the equals comparison.</returns>
-    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    /// <returns>The converted value.</returns>
+    public override object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-        return Equals(value, parameter) ? IsEqual : !IsEqual;
+        return Equals(value, CompareWith) ? IsEqual : IsNotEqual;
     }
 
     /// <summary>