+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added DoubleComparisonToVisibilityConverter to execute a comparison on a single Visibility or list of Visibilities to a boolean representation.
 - Added DoubleMathConverter to do a simple calculation with the given value as double.
 - Added DoubleToStringConverter to format a double to string.
+### Changed
+- Move the parameter for the EqualsToBooleanConverter to a converter property.
+- Change the configurable return value for the EqualsToBooleanConverter from a single to multiple properties.
 ### Removed
 - Removed InverseBooleanConverter, thats replaced by BooleanToBooleanConverter.
 - Removed MultiBooleanToBooleanConverter, thats integrated into the BooleanToBooleanConverter.